Added: app - fixed step loop as can be seen here: https://gafferongames.com/post/fix_your_timestep/ Added: glfw - virtual buttons that can map from any kind of input (or combination of inputs) to an actionevery button that have native support for callback knows exactly how many transition states there are (from pushed to released)

+++ system.h
@@ -5,8 +5,8 @@
 
 void system__init();
 
-void system__sleep(uint32_t s);
-void system__usleep(uint64_t us);
+void system__sleep(double s);
+void system__usleep(double us);
 
 // @returns returns time in seconds
 double system__get_time();
